requirements.txt
spacy==3.7.2
ru-core-news-sm==3.7.0
telethon==1.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
numpy==1.24.3
scikit-learn==1.3.0
python-Levenshtein==0.21.1
fuzzywuzzy==0.18.0 
openai